Use raw os to make things easier
